+++ LightBlue/LightBlueContext.cs
@@ -1,20 +1,33 @@
 ﻿using System;
 
+using LightBlue.External;
 using LightBlue.Hosted;
 using LightBlue.Setup;
 using LightBlue.Standalone;
+
+using Microsoft.WindowsAzure.ServiceRuntime;
 
 namespace LightBlue
 {
     public static class LightBlueContext
     {
         private static bool _initialised;
+        private static string _roleName;
         private static IAzureSettings _azureSettings;
         private static IAzureLocalResourceSource _azureLocalResources;
         private static Func<string, IAzureStorage> _azureStorageFactory;
         private static Func<Uri, IAzureBlobContainer> _azureBlobContainerFactory;
         private static Func<Uri, IAzureBlockBlob> _azureBlockBlobFactory;
         private static AzureEnvironment _azureEnvironment;
+
+        public static string RoleName
+        {
+            get
+            {
+                Initialise();
+                return _roleName;
+            }
+        }
 
         public static IAzureSettings AzureSettings
         {
@@ -77,35 +90,64 @@ namespace LightBlue
                 return;
             }
 
-            _azureEnvironment = LightBlueConfiguration.DetermineEnvironment();
+            var environmentDefinition = LightBlueConfiguration.DetermineEnvironmentDefinition();
+
+            _azureEnvironment = environmentDefinition.AzureEnvironment;
 
             if (_azureEnvironment == AzureEnvironment.LightBlue)
             {
-                InitialiseAsLightBlue();
+                InitialiseAsLightBlue(environmentDefinition);
             }
             else
             {
-                InitialiseAsHosted();
+                InitialiseAsHosted(environmentDefinition);
             }
             _initialised = true;
         }
 
-        private static void InitialiseAsLightBlue()
+        private static void InitialiseAsLightBlue(EnvironmentDefinition environmentDefinition)
         {
-            _azureSettings = new StandaloneAzureSettings(LightBlueConfiguration.StandaloneConfiguration);
-            _azureLocalResources = new StandaloneAzureLocalResourceSource(LightBlueConfiguration.StandaloneConfiguration, StandaloneEnvironment.LightBlueDataDirectory);
+            if (environmentDefinition.HostingType == HostingType.Role)
+            {
+                _roleName = environmentDefinition.StandaloneConfiguration.RoleName;
+                _azureSettings = new StandaloneAzureSettings(environmentDefinition.StandaloneConfiguration);
+                _azureLocalResources = new StandaloneAzureLocalResourceSource(
+                    environmentDefinition.StandaloneConfiguration,
+                    StandaloneEnvironment.LightBlueDataDirectory);
+            }
+            else
+            {
+                ConfigureElementsNotAvailableExternalToHost();
+            }
+
             _azureStorageFactory = connectionString => new StandaloneAzureStorage(connectionString);
             _azureBlobContainerFactory = uri => new StandaloneAzureBlobContainer(uri);
             _azureBlockBlobFactory = blobUri => new StandaloneAzureBlockBlob(blobUri);
         }
 
-        private static void InitialiseAsHosted()
+        private static void InitialiseAsHosted(EnvironmentDefinition environmentDefinition)
         {
-            _azureSettings = new HostedAzureSettings();
-            _azureLocalResources = new HostedAzureLocalResourceSource();
+            if (environmentDefinition.HostingType == HostingType.Role)
+            {
+                _roleName = RoleEnvironment.CurrentRoleInstance.Role.Name;
+                _azureSettings = new HostedAzureSettings();
+                _azureLocalResources = new HostedAzureLocalResourceSource();
+            }
+            else
+            {
+                ConfigureElementsNotAvailableExternalToHost();
+            }
+
             _azureStorageFactory = connectionString => new HostedAzureStorage(connectionString);
             _azureBlobContainerFactory = uri => new HostedAzureBlobContainer(uri);
             _azureBlockBlobFactory = blobUri => new HostedAzureBlockBlob(blobUri);
+        }
+
+        private static void ConfigureElementsNotAvailableExternalToHost()
+        {
+            _roleName = "External";
+            _azureSettings = new ExternalAzureSettings();
+            _azureLocalResources = new ExternalAzureLocalResourceSource();
         }
     }
 }