+++ LightBlue/Setup/LightBlueHostType.cs
@@ -1,0 +1,8 @@
+﻿namespace LightBlue.Setup
+{
+    public enum LightBlueHostType
+    {
+        Direct = 0,
+        Indirect = 1
+    }
+}